attribute macro for malloc/calloc/free injection
