+++ Cargo.toml
@@ -32,6 +32,7 @@ no-threads = ["blst/no-threads"]
 blst = { version = "0.3.11", default-features = false, features = [ "no-threads" ] }
 hex = { version = "0.4.2", default-features = false, features = ["alloc"] }
 libc = { version = "0.2", default-features = false }
+c-kzg-risc0-alloc = { path = "./c-kzg-risc0-alloc" }
 serde = { version = "1.0", optional = true, default-features = false, features = [
     "alloc",
     "derive",
