+++ bindings/rust/src/lib.rs
@@ -22,63 +22,5 @@ pub use bindings::{
 };
 // Expose the remaining relevant types.
 pub use bindings::{Blob, Bytes32, Bytes48, Error};
-
-#[cfg(target_os = "zkvm")]
-pub mod alloc_builtin {
-    extern crate alloc;
-
-    const LEN_SIZE: usize = core::mem::size_of::<usize>();
-
-    unsafe fn store_len(block_ptr: *mut u8, size: usize) -> *mut u8 {
-        let len_ptr = block_ptr as *mut usize;
-        *len_ptr = size;
-
-        block_ptr.offset(LEN_SIZE as isize)
-    }
-
-    unsafe fn get_len(block_ptr: *mut u8) -> (usize, *mut u8) {
-        let len_ptr = block_ptr.offset(-(LEN_SIZE as isize)) as *mut usize;
-        (*len_ptr, len_ptr as *mut u8)
-    }
-
-    #[no_mangle]
-    unsafe extern "C" fn malloc(size: usize) -> *mut core::ffi::c_void {
-        let layout = core::alloc::Layout::from_size_align(size + LEN_SIZE, 4)
-            .expect("unable to construct memory layout");
-        let block_ptr = alloc::alloc::alloc(layout);
-
-        if block_ptr.is_null() {
-            alloc::alloc::handle_alloc_error(layout);
-        }
-
-        let data_ptr = store_len(block_ptr, size);
-
-        data_ptr as *mut core::ffi::c_void
-    }
-
-    #[no_mangle]
-    unsafe extern "C" fn calloc(nobj: usize, size: usize) -> *mut core::ffi::c_void {
-        let size = nobj * size;
-        let layout = core::alloc::Layout::from_size_align(size + LEN_SIZE, 4)
-            .expect("unable to construct memory layout");
-        let block_ptr = alloc::alloc::alloc_zeroed(layout);
-
-        if block_ptr.is_null() {
-            alloc::alloc::handle_alloc_error(layout);
-        }
-
-        let data_ptr = store_len(block_ptr, size);
-
-        data_ptr as *mut core::ffi::c_void
-    }
-
-    #[no_mangle]
-    unsafe extern "C" fn free(block_ptr: *const core::ffi::c_void) {
-        let (size, free_ptr) = get_len(block_ptr as *mut u8);
-
-        let layout = core::alloc::Layout::from_size_align(size, 4)
-            .expect("unable to construct memory layout");
-
-        alloc::alloc::dealloc(free_ptr, layout);
-    }
-}
+// Expose the attribute macro to generate malloc/calloc/free definitions
+pub use c_kzg_risc0_alloc::risc0_c_kzg_alloc_mod;
