+++ c-kzg-risc0-alloc/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "c-kzg-risc0-alloc"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "2.0.72"
+quote = "1.0.36"
